Move Connection display components into a separate namespace
