+++ src/ConnectedMode/ConnectedMode.csproj
@@ -56,7 +56,8 @@
 
   <ItemGroup>
     <Page Include="Migration\Wizard\MigrationWizardWindow.xaml" />
-    <Page Include="UI\ConnectionIconComponent.xaml" />
+    <Page Include="UI\ConnectionDisplay\ConnectionIconComponent.xaml" />
+    <Page Include="UI\ConnectionDisplay\ConnectionInfoComponent.xaml" />
     <Page Include="UI\ProgressAndErrorReporterComponent.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -79,7 +80,6 @@
     </Page>
     <Page Include="UI\ServerSelection\ServerSelectionDialog.xaml" />
     <Page Include="UI\ManageConnections\ManageConnectionsDialog.xaml" />
-    <Page Include="UI\ConnectionInfoComponent.xaml" />
     <Page Include="UI\OrganizationSelection\ManualOrganizationSelectionDialog.xaml" />
     <Page Include="UI\OrganizationSelection\OrganizationSelectionDialog.xaml" />
     <Page Include="UI\ProjectSelection\ProjectSelectionDialog.xaml" />
