+++ src/ConnectedMode/UI/ConnectionDisplay/ConnectionIconComponent.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="SonarLint.VisualStudio.ConnectedMode.UI.ConnectionIconComponent"
+﻿<UserControl x:Class="SonarLint.VisualStudio.ConnectedMode.UI.ConnectionDisplay.ConnectionIconComponent"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="d"
