+++ src/ConnectedMode/UI/Credentials/CredentialsDialog.xaml
@@ -7,6 +7,7 @@
         WindowStartupLocation="CenterOwner"
         xmlns:vsimagecatalog="clr-namespace:Microsoft.VisualStudio.Imaging;assembly=Microsoft.VisualStudio.ImageCatalog"
         xmlns:imaging="clr-namespace:Microsoft.VisualStudio.Imaging;assembly=Microsoft.VisualStudio.Imaging"
+        xmlns:connectionDisplay="clr-namespace:SonarLint.VisualStudio.ConnectedMode.UI.ConnectionDisplay"
         x:Name="This">
     <Window.Resources>
         <ResourceDictionary>
@@ -32,7 +33,7 @@
         <TextBlock Grid.Row="0" VerticalAlignment="Center" Margin="0, 10, 0, 0">
             <Run Text="{x:Static res:UiResources.AuthenticationDescription}" />
             <InlineUIContainer Style="{StaticResource ConnectionInfoInlineWrapper}">
-                <ui:ConnectionInfoComponent ConnectionInfo="{Binding Path=ConnectionInfo}"
+                <connectionDisplay:ConnectionInfoComponent ConnectionInfo="{Binding Path=ConnectionInfo}"
                                             TextAndIconVerticalAlignment="Center" Margin="5,0" />
             </InlineUIContainer>
         </TextBlock>
